requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
netCDF4==1.6.5
xarray==2023.12.0
h5py==3.10.0


